# public/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Next Swagger API Example",
    "version": "1.0"
  },
  "paths": {
...
    "/api/group-event/{id}/join/{personId}": {
      "put": {
        "summary": "Update participation options for a person in a group event",
        "description": "Updates the possible options for a person's participation in a specific group event.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the group event"
          },
          {
            "in": "path",
            "name": "personId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the person (invitee)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "possibleOptions"
                ],
                "properties": {
                  "possibleOptions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Array of possible option IDs that the person is interested in"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Participation options updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "personId",
                    "possibleOptionIds"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Identifier of the updated invitee"
                    },
                    "personId": {
                      "type": "string",
                      "description": "Identifier of the person (invitee)"
                    },
                    "possibleOptionIds": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "Updated array of possible option IDs for the invitee"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, object invalid or invitee not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/group-event/{id}/join": {
      "post": {
        "summary": "Join a group event",
        "description": "Allows a person to join a group event using the event ID and person's name.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the group event to join"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "person"
                ],
                "properties": {
                  "person": {
                    "type": "object",
                    "description": "The person joining the event",
                    "required": [
                      "name"
                    ],
                    "properties": {
                      "name": {
                        "type": "string",
                        "description": "Name of the person"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Person joined the event successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Id of the person who joined"
                    },
                    "name": {
                      "type": "string",
                      "description": "Name of the person who joined"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, object invalid"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}